Task 5   - clean up   - add changelog   - default Ui._UiCommandFactory will be for current iteration. not required to set custom UiCommandFactory after Ui object is created
