+++ docs/CHANGELOG.md
@@ -41,3 +41,8 @@ Add the ability to store whatever text entered by the user and display them back
 `done [id]` set task as done \
 
 ![level 4](/docs/screenshots/branch-level-4.gif)
+
+
+## Feature Level 5
+
+No feature changes (error handling)
