--- src/main/java/duke/command/systemCommand/CommandEcho.java
@@ -1,12 +1,0 @@
-package duke.command.systemCommand;
-
-import duke.command.Command;
-import duke.dukeUtility.enums.ResponseType;
-
-import java.util.List;
-
-public class CommandEcho extends Command {
-    public CommandEcho(String text) {
-        super(ResponseType.ECHO, List.of("echo", text));
-    }
-}
