+++ src/main/java/duke/command/taskCommand/taskAdd/CommandAddNewEvent.java
@@ -4,7 +4,6 @@ import duke.command.Command;
 import duke.dukeUtility.enums.ResponseType;
 import duke.TaskManager;
 
-import java.time.LocalDateTime;
 import java.util.List;
